Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -3,7 +3,7 @@
 int main(){
     int a;
     scanf("%d", &a);
-    int b = 1<<31;
-    printf("%s", (a & b)? "Set" : "Not Set"); 
+    unsigned int b = 1U<<31;
+    printf("%s", ((unsigned int)a & b)? "Set" : "Not Set"); 
     return 0;
 }
